+++ services/post.service.js
@@ -1,13 +1,22 @@
 const postDb = require('../db/post.db');
 const commentsService = require('./comments.service');
 const imageService = require('./images.service');
+const { errorFactory, errors } = require('../utils/errorManager');
+
+const doesPostExist = async (postId, userId) => {
+  const post = await postDb.getPostById(postId, userId);
+  return !!post;
+};
 
 const changeLikedByToBool = (post) => ({ ...post, likedByUser: !!post.likedByUser });
 
 const getPostById = async (postId, userId) => {
   const post = await postDb.getPostById(postId, userId);
-  post.comments = await commentsService.getCommentsByPostId(postId, userId);
-  return changeLikedByToBool(post);
+  if (post) {
+    post.comments = await commentsService.getCommentsByPostId(postId, userId);
+    return changeLikedByToBool(post);
+  }
+  throw errorFactory(errors.postDoesntExist, 'a post with the specified id ws not found');
 };
 
 const createPost = async (post, imageBuffer) => {
@@ -24,20 +33,27 @@ const createPost = async (post, imageBuffer) => {
 
 const getPosts = async (postFilters, userId) => {
   const posts = await postDb.getPosts(postFilters, userId);
-
-  const formattedPosts = posts.map(changeLikedByToBool);
-
-  return formattedPosts;
+  if (posts) {
+    const formattedPosts = posts.map(changeLikedByToBool);
+    return formattedPosts;
+  }
+  return [];
 };
 
+
 const addPostLike = async (postId, userId) => {
+  if (!doesPostExist(postId, userId)) {
+    throw errorFactory(errors.postDoesntExist, 'the post to like doesnt exist');
+  }
   await postDb.addPostLike(postId, userId);
 };
 
 const deletePostLike = async (postId, userId) => {
+  if (!doesPostExist(postId, userId)) {
+    throw errorFactory(errors.postDoesntExist, 'the post to like doesnt exist');
+  }
   await postDb.deletePostLike(postId, userId);
 };
-
 
 module.exports = {
   getPostById, createPost, getPosts, addPostLike, deletePostLike,
